(PC-11330) [review] move to dedicated module
