+++ api/src/pcapi/core/users/api.py
@@ -18,7 +18,6 @@ from flask import current_app as app
 from flask_jwt_extended import create_access_token
 from google.cloud.storage.blob import Blob
 from redis import Redis
-from sentry_sdk import capture_exception
 
 # TODO (viconnex): fix circular import of pcapi/models/__init__.py
 from pcapi import models  # pylint: disable=unused-import
@@ -29,16 +28,14 @@ import pcapi.core.bookings.repository as bookings_repository
 import pcapi.core.fraud.api as fraud_api
 import pcapi.core.fraud.models as fraud_models
 from pcapi.core.mails.transactional import users as user_emails
-from pcapi.core.mails.transactional.users.email_address_change import send_confirmation_email_change_email
-from pcapi.core.mails.transactional.users.email_address_change import send_information_email_change_email
 from pcapi.core.mails.transactional.users.email_confirmation_email import send_email_confirmation_email
 import pcapi.core.payments.api as payment_api
 from pcapi.core.subscription import api as subscription_api
 from pcapi.core.subscription import messages as subscription_messages
 from pcapi.core.subscription.models import BeneficiaryPreSubscription
 import pcapi.core.subscription.repository as subscription_repository
+from pcapi.core.users import email as email_api
 from pcapi.core.users import exceptions
-from pcapi.core.users import repository as users_repository
 from pcapi.core.users.external import update_external_user
 from pcapi.core.users.models import Credit
 from pcapi.core.users.models import DomainsCredit
@@ -52,7 +49,6 @@ from pcapi.core.users.models import UserRole
 from pcapi.core.users.models import VOID_PUBLIC_NAME
 from pcapi.core.users.repository import does_validated_phone_exist
 from pcapi.core.users.utils import delete_object
-from pcapi.core.users.utils import encode_jwt_payload
 from pcapi.core.users.utils import get_object
 from pcapi.core.users.utils import sanitize_email
 from pcapi.core.users.utils import store_object
@@ -81,7 +77,6 @@ from pcapi.tasks.account import verify_identity_document
 from pcapi.utils import phone_number as phone_number_utils
 from pcapi.utils.sentry import suppress_and_capture_errors
 from pcapi.utils.token import random_token
-from pcapi.utils.urls import generate_firebase_dynamic_link
 
 
 logger = logging.getLogger(__name__)
@@ -502,16 +497,6 @@ def bulk_unsuspend_account(user_ids: list[int], actor: User) -> None:
     )
 
 
-def send_user_emails_for_email_change(user: User, new_email: str, expiration_date: datetime) -> None:
-    user_with_new_email = find_user_by_email(new_email)
-    if user_with_new_email:
-        return
-
-    send_information_email_change_email(user)
-    link_for_email_change = _build_link_for_email_change(user.email, new_email, expiration_date)
-    send_confirmation_email_change_email(user, new_email, link_for_email_change)
-
-
 def change_user_email(current_email: str, new_email: str) -> None:
     current_user = user_queries.find_user_by_email(current_email)
     if not current_user:
@@ -542,7 +527,7 @@ def change_user_email(current_email: str, new_email: str) -> None:
         # It is not a huge problem if an error is encountered, no need
         # to raise a 500 error and alert the end user, as long as it
         # is sent to sentry.
-        app.redis_client.unlink(email_update_token_ttl_key(current_user))
+        app.redis_client.unlink(email_api.email_update_token_ttl_key(current_user))
 
 
 def update_user_info(
@@ -576,16 +561,6 @@ def update_user_info(
     if public_name is not UNCHANGED:
         user.publicName = public_name
     repository.save(user)
-
-
-def _build_link_for_email_change(current_email: str, new_email: str, expiration_date: datetime) -> str:
-    token = encode_jwt_payload({"current_email": current_email, "new_email": new_email}, expiration_date)
-    expiration = int(expiration_date.timestamp())
-
-    path = "changement-email"
-    params = {"token": token, "expiration_timestamp": expiration}
-
-    return generate_firebase_dynamic_link(path, params)
 
 
 def get_domains_credit(user: User, user_bookings: list[bookings_models.Booking] = None) -> Optional[DomainsCredit]:
@@ -826,64 +801,6 @@ def check_and_update_phone_validation_attempts(redis: Redis, user: User) -> None
         redis.expire(phone_validation_attempts_key, settings.PHONE_VALIDATION_ATTEMPTS_TTL)
 
 
-def check_email_update_attempts(user: User, redis: Redis) -> None:
-    update_email_attempts_key = f"update_email_attemps_user_{user.id}"
-    count = redis.incr(update_email_attempts_key)
-
-    if count == 1:
-        redis.expire(update_email_attempts_key, settings.EMAIL_UPDATE_ATTEMPTS_TTL)
-
-    if count > settings.MAX_EMAIL_UPDATE_ATTEMPTS:
-        raise exceptions.EmailUpdateLimitReached()
-
-
-def email_update_token_ttl_key(user: User) -> str:
-    return f"update_email_active_tokens_{user.id}"
-
-
-def save_email_update_activation_token_counter(user: User, redis: Redis) -> datetime:
-    """
-    Use a dummy counter to find out whether the user already has an
-    active token.
-
-    * If the incr command returns 1, there were none. Hence, set a TTL
-      (expiration_date, the lifetime of the validation token).
-    * If not, raise an error because there is already one.
-    """
-    key = email_update_token_ttl_key(user)
-    count = redis.incr(key)
-
-    if count > 1:
-        raise exceptions.EmailUpdateTokenExists()
-
-    expiration_date = datetime.now() + constants.EMAIL_CHANGE_TOKEN_LIFE_TIME
-    redis.expireat(key, expiration_date)
-
-    return expiration_date
-
-
-def check_user_password(user: User, password: Optional[str]) -> None:
-    if not password:
-        raise exceptions.EmailUpdateInvalidPassword()
-
-    try:
-        users_repository.check_user_and_credentials(user, password)
-    except exceptions.InvalidIdentifier as exc:
-        raise exceptions.EmailUpdateInvalidPassword() from exc
-    except exceptions.UnvalidatedAccount as exc:
-        # This should not happen. But, if it did:
-        # 1. send the error to sentry
-        # 2. raise the same error as above, so the end client
-        # can't guess what happened.
-        capture_exception(exc)
-        raise exceptions.EmailUpdateInvalidPassword() from exc
-
-
-def check_email_address_does_not_exist(email: str) -> None:
-    if user_queries.find_user_by_email(email):
-        raise exceptions.EmailExistsError(email)
-
-
 def check_phone_number_not_used(phone_number: str) -> None:
     if does_validated_phone_exist(phone_number):
         raise exceptions.PhoneAlreadyExists(phone_number)
@@ -1042,15 +959,6 @@ def update_user_profile(user: User, content: "account_serialization.UserProfileU
     if content.subscriptions is not None:
         update_notification_subscription(user, content.subscriptions)
         update_external_user(user)
-
-
-def update_email(user: User, email: str, password: Optional[str]) -> None:
-    check_email_update_attempts(user, app.redis_client)
-    check_email_address_does_not_exist(email)
-    check_user_password(user, password)
-
-    expiration_date = save_email_update_activation_token_counter(user, app.redis_client)
-    send_user_emails_for_email_change(user, email, expiration_date)
 
 
 def update_notification_subscription(
