+++ src/pcapi/core/users/email/__init__.py
@@ -1,0 +1,4 @@
+from .send import send_user_emails_for_email_change  # noqa: F401
+from .update import email_update_token_ttl_key  # noqa: F401
+from .update import save_email_update_activation_token_counter  # noqa: F401
+from .update import update_email  # noqa: F401
